making sure schema.js just contains schema-related stuff.
