+++ after/src/schema.js
@@ -6,35 +6,6 @@ const {
   GraphQLNonNull
 } = require('graphql')
 
-const getKeyValue = db => key => {
-  const params = {
-    TableName: process.env.DB_TABLE,
-    Key: { key }
-  }
-
-  return new Promise((resolve, reject) => {
-    db.get(params, (error, data) => {
-      if (error) {
-        reject(error)
-      } else {
-        resolve(typeof data.Item === 'undefined' ? '' : data.Item.info)
-      }
-    })
-  })
-}
-
-const setKeyValue = db => (key, info) => {
-  const params = {
-    TableName: process.env.DB_TABLE,
-    Key: { key: key },
-    UpdateExpression: 'SET info = :info',
-    ExpressionAttributeValues: {
-      ':info': info
-    }
-  }
-  return db.update(params).promise().then(() => info)
-}
-
 module.exports = db => new GraphQLSchema({
   mutation: new GraphQLObjectType({
     name: 'RootMutationType',
@@ -45,7 +16,7 @@ module.exports = db => new GraphQLSchema({
           value: { name: 'value', type: new GraphQLNonNull(GraphQLString) }
         },
         type: GraphQLString,
-        resolve: (parent, args) => setKeyValue(db)(args.key, args.value)
+        resolve: (parent, args) => db.setKeyValue(args.key, args.value)
       }
     }
   }),
@@ -55,7 +26,7 @@ module.exports = db => new GraphQLSchema({
       getKey: {
         args: { key: { name: 'key', type: new GraphQLNonNull(GraphQLString) } },
         type: GraphQLString,
-        resolve: (parent, args) => getKeyValue(db)(args.key)
+        resolve: (parent, args) => db.getKeyValue(args.key)
       }
     }
   })
